Speed ups: Stub.out is now a property and Chunk.data is now joined only when necessary.
Note that Chunk.data is *not* a property.  Because `flushable' guards
access to Chunk.data, the underlying list only needs to be joined when
that is set to True.  The results of Stub.out, however, can be asked
for in media res (see AsyncKey in tests.dust_site) so a property is
appropriate here.